initial version

+++ Projeto2/controllers/deletarController.php
@@ -1,8 +1,19 @@
 <?php
 include_once "connect.php";
 
+$deletar = addslashes($_POST['deletar']);
 $iddeletar = addslashes($_POST['id']);
 
-mysqli_query($link, "delete from cliente where cliente_id = '$iddeletar'");
+if ($deletar == 0)
+    {
+        mysqli_query($link, "delete from cliente where cliente_id = '$iddeletar'");
 
-header('location:../deletar.php');
+        header("Refresh:0");
+    }
+
+else if($deletar == 1)
+{
+mysqli_query($link, "delete from endereco where endereco_id = '$iddeletar'");
+
+header("Refresh:0");
+}
